[test/dynamo] inline misc and repro test splits
