--- test/dynamo/dynamo_test_common.py
@@ -1,190 +1,0 @@
-# Owner(s): ["module: dynamo"]
-# flake8: noqa: B001,B006,B020,B021,B950,C405,C416,E711,E721,E722,E731,F401,F403,F405,F541,F821,F823
-# ruff: noqa: F401,F541
-import abc
-import builtins
-import collections
-import collections.abc
-import copy
-import dataclasses
-import dis
-import enum
-import functools
-import gc
-import importlib
-import itertools
-import json
-import logging
-import math
-import operator
-import os
-import pickle
-import random
-import re
-import sys
-import tempfile
-import threading
-import traceback
-import types
-import typing
-import unittest
-import unittest.mock as mock
-import warnings
-import weakref
-from unittest.mock import patch
-
-import numpy as np
-
-import torch
-import torch._dynamo.testing
-import torch._inductor.config
-import torch._inductor.test_case
-import torch.onnx.operators
-import torch.utils._pytree as python_pytree
-import torch.utils.cpp_extension
-from torch import Tensor
-from torch._C import FileCheck
-from torch._dynamo import allow_in_graph
-from torch._dynamo.comptime import comptime
-from torch._dynamo.eval_frame import _debug_get_cache_entry_list
-from torch._dynamo.exc import Unsupported
-from torch._dynamo.source import ConstantSource, GetItemSource, LocalSource
-from torch._dynamo.testing import (
-    CompileCounter,
-    CompileCounterWithBackend,
-    expectedFailureDynamic,
-    same,
-    skipIfNotPy311,
-    unsupported,
-)
-from torch._dynamo.utils import call_size, counters, ifdynstaticdefault
-from torch._dynamo.variables import builder
-from torch._inductor.codecache import WritableTempFile
-from torch._inductor.utils import fresh_cache, run_and_get_code
-from torch.ao.quantization import MinMaxObserver
-from torch.ao.quantization.fake_quantize import FakeQuantize
-from torch.ao.quantization.qconfig import QConfig
-from torch.ao.quantization.quantize_fx import prepare_qat_fx
-from torch.fx.experimental.recording import NotEqualError, replay_shape_env_events
-from torch.fx.experimental.symbolic_shapes import (
-    _constrain_range_for_size,
-    constrain_range,
-    constrain_unify,
-    ConstraintViolationError,
-    expect_true,
-    guard_or_false,
-    guard_size_oblivious,
-    ShapeEnv,
-)
-from torch.nn import functional as F
-from torch.testing import make_tensor
-from torch.testing._internal.common_cuda import (
-    PLATFORM_SUPPORTS_FLASH_ATTENTION,
-    SM80OrLater,
-    TEST_CUDA,
-)
-from torch.testing._internal.common_device_type import instantiate_device_type_tests
-from torch.testing._internal.common_methods_invocations import (
-    sample_inputs_take_along_dim,
-)
-from torch.testing._internal.common_utils import (
-    freeze_rng_state,
-    instantiate_parametrized_tests,
-    IS_FBCODE,
-    parametrize,
-    scoped_load_inline,
-    set_default_dtype,
-    skipIfHpu,
-    skipIfNNModuleInlined,
-    skipIfWindows,
-    subtest,
-    TEST_HPU,
-    TEST_XPU,
-    wrapDeterministicFlagAPITest,
-)
-from torch.testing._internal.jit_utils import JitTestCase
-
-
-pytree_modules = {
-    "python": python_pytree,
-}
-if python_pytree._cxx_pytree_dynamo_traceable:
-    import torch.utils._cxx_pytree as cxx_pytree
-
-    pytree_modules["cxx"] = cxx_pytree
-    pytree_modules["native_optree"] = cxx_pytree.optree
-else:
-    cxx_pytree = None
-
-parametrize_pytree_module = parametrize(
-    "pytree",
-    [subtest(module, name=name) for name, module in pytree_modules.items()],
-)
-
-MyTuple = collections.namedtuple("MyTuple", ["a", "b", "ab"])
-T = typing.TypeVar("T")
-
-
-# Defined in CPython's Include/object.h
-TPFLAGS_MAPPING = 1 << 6
-
-GLOBAL_INT = 1
-
-
-# Specializes a test to run only if translation validation is set.
-def onlyIfTranslationValidation(fn: typing.Callable) -> typing.Callable:
-    @functools.wraps(fn)
-    def wrapper(*args, **kwargs):
-        import torch.fx.experimental.validator
-
-        if torch.fx.experimental.validator.translation_validation_enabled():
-            return fn(*args, **kwargs)
-        raise unittest.SkipTest(f"only works when TV is True.")
-
-    return wrapper
-
-
-class MyPickledModule(torch.nn.Module):
-    def __init__(self, z):
-        super().__init__()
-        self.z = z
-
-    def forward(self, x, y):
-        return x * x * x + y + self.z
-
-
-# These are used for test_{cond/map}_with_quantization
-default_symmetric_fake_quant = FakeQuantize.with_args(
-    observer=MinMaxObserver, qscheme=torch.per_tensor_symmetric, dtype=torch.quint8
-)
-default_weight_symmetric_fake_quant = FakeQuantize.with_args(
-    observer=MinMaxObserver, qscheme=torch.per_tensor_symmetric, dtype=torch.qint8
-)
-uniform_qconfig_8bit = QConfig(
-    activation=default_symmetric_fake_quant,
-    weight=default_weight_symmetric_fake_quant.with_args,
-)
-qconfig_dict = {"object_type": [(torch.nn.Linear, uniform_qconfig_8bit)]}
-
-
-def closure_adder(val):
-    def inner(x):
-        return torch.sin(x + val)
-
-    return inner
-
-
-class UserDefineSetAttr:
-    setup = False
-
-    def __setattr__(self, key, value):
-        assert torch.compiler.is_dynamo_compiling() or UserDefineSetAttr.setup  # noqa: S101
-        super().__setattr__(f"pfx_{key}", value)
-
-    def __getattr__(self, key, c=1):
-        assert torch.compiler.is_dynamo_compiling() or UserDefineSetAttr.setup  # noqa: S101
-        # c is added to force a guard on __defaults__ and checks the source for __getattr__
-        if c:
-            return self.__dict__[f"pfx_{key}"]
-        else:
-            return None
