+++ test/dynamo/test_device.py
@@ -1,16 +1,22 @@
 # Owner(s): ["module: dynamo"]
-# flake8: noqa: B001,B006,B020,B021,B950,C405,C416,E711,E721,E722,E731,F401,F403,F405,F541,F821,F823
-# ruff: noqa: B020,F403,F405,F841,PLW0127
-try:
-    from .dynamo_test_common import *
-except ImportError:
-    from dynamo_test_common import *
+import unittest
+from unittest import mock
+
+import torch
+import torch._dynamo.testing
+import torch._functorch.config
+import torch._inductor.test_case
+from torch._dynamo.testing import same
+from torch.testing._internal.common_cuda import TEST_CUDA
+from torch.testing._internal.common_device_type import instantiate_device_type_tests
+
+
+requires_cuda = unittest.skipUnless(torch.cuda.is_available(), "requires cuda")
 
 
 class MiscTestsDevice(torch._inductor.test_case.TestCase):
     def test_rand(self, device):
         cnts = torch._dynamo.testing.CompileCounter()
-        device = device
 
         def fn():
             return torch.randn(10, device=device)
@@ -27,57 +33,6 @@ class MiscTestsDevice(torch._inductor.test_case.TestCase):
         res = opt_fn()
 
         self.assertTrue(same(res, ref_run1))
-
-    @unittest.skipIf(
-        not PLATFORM_SUPPORTS_FLASH_ATTENTION,
-        "Can't run fused SDPA on this platform",
-    )
-    def test_parsing_sdpa(self, device):
-        class MyModule(torch.nn.Module):
-            def forward(self, query, key, value):
-                out = F.scaled_dot_product_attention(query, key, value, None, 0, True)
-                out = F.scaled_dot_product_attention(
-                    query, key, value, None, 0, True, scale=8
-                )
-                out = F.scaled_dot_product_attention(
-                    query=query,
-                    key=key,
-                    value=value,
-                    attn_mask=None,
-                    dropout_p=0,
-                    is_causal=True,
-                )
-                out = F.scaled_dot_product_attention(
-                    query,
-                    key=key,
-                    value=value,
-                    attn_mask=None,
-                    dropout_p=0,
-                    is_causal=True,
-                )
-                out = F.scaled_dot_product_attention(
-                    query, key, value, None, dropout_p=0, is_causal=True
-                )
-                out = F.scaled_dot_product_attention(query, key, value, None, scale=8)
-                return out
-
-        device = device
-        dtype = torch.float16
-        seq_len_q = 1
-        seq_len_k = 1
-        head_dim = 8
-        query = torch.ones(
-            1, 8, seq_len_q, head_dim, device=device, dtype=dtype, requires_grad=True
-        )
-        key = torch.ones(
-            1, 8, seq_len_k, head_dim, device=device, dtype=dtype, requires_grad=True
-        )
-        value = torch.ones(
-            1, 8, seq_len_k, head_dim, device=device, dtype=dtype, requires_grad=True
-        )
-        module = MyModule()
-        opt_mod = torch.compile(module, backend="inductor")
-        opt_mod(query, key, value)
 
     def test_torch_device_is_available(self, device):
         def fn(x):
@@ -124,14 +79,14 @@ class MiscTestsDevice(torch._inductor.test_case.TestCase):
         ):
             x1 = torch.rand(4).to(device)
             opt_fn1 = torch.compile(fn1, backend="eager", fullgraph=True)
-            res1 = opt_fn1(x1)
+            opt_fn1(x1)
 
         with self.assertRaisesRegex(
             AssertionError, "Expect 1 input to cudnn.is_acceptable"
         ):
             x2 = torch.rand(4).to(device)
             opt_fn2 = torch.compile(fn2, backend="eager", fullgraph=True)
-            res = opt_fn2(x2)
+            opt_fn2(x2)
 
     @unittest.skipIf(not torch.cuda.is_available(), "requires cuda")
     @torch._dynamo.config.patch(recompile_limit=999)
@@ -181,10 +136,10 @@ class MiscTestsDevice(torch._inductor.test_case.TestCase):
         self.assertEqual(out, opt_out)
 
     def test_torch_device_python_type(self, device):
-        device_type = torch.device(device).type
-        for device, device_type, index in [
+        default_device_type = torch.device(device).type
+        for current_device, current_device_type, index in [
             ("cpu", "cpu", None),
-            (device, device_type, 0),
+            (device, default_device_type, 0),
         ]:
 
             def fn(target):
@@ -192,7 +147,7 @@ class MiscTestsDevice(torch._inductor.test_case.TestCase):
                 a = torch.zeros(2, 3, device=target_device)
                 # Constant assert at trace time
                 assert isinstance(target_device, torch.device)  # noqa: S101
-                assert target_device.type == device_type  # noqa: S101
+                assert target_device.type == current_device_type  # noqa: S101
                 assert target_device.index == index  # noqa: S101
                 b = torch.zeros(2, 3, device=target_device)
                 c = torch.zeros(2, 3, device=target_device)
@@ -200,12 +155,12 @@ class MiscTestsDevice(torch._inductor.test_case.TestCase):
 
             from torch._dynamo.variables import ConstantVariable
 
-            device = torch.device(device)
-            expected_variable = ConstantVariable(device)
-            self.assertEqual(expected_variable.python_type(), type(device))
+            current_device = torch.device(current_device)
+            expected_variable = ConstantVariable(current_device)
+            self.assertEqual(expected_variable.python_type(), type(current_device))
 
             opt_func = torch.compile(fn, backend="eager", fullgraph=True)
-            a = torch.tensor([2, 3], device=device)
+            a = torch.tensor([2, 3], device=current_device)
             res = opt_func(a)
             self.assertIsInstance(res, torch.Tensor)
 
@@ -297,16 +252,99 @@ class MiscTestsDevice(torch._inductor.test_case.TestCase):
 
         foo = Foo()
         inputs = torch.randn(3, 4)
-        result = foo.forward(inputs)
+        foo.forward(inputs)
 
-        original_pad_val = foo.config.pad_val
         foo.config.pad_val += 1.0
-        result2 = foo.forward(inputs)
+        foo.forward(inputs)
+
+
+class ReproDeviceRuntimeTests(torch._inductor.test_case.TestCase):
+    def test_guard_default_device(self, device):
+        try:
+            torch.set_default_device(device)
+
+            counter = torch._dynamo.testing.CompileCounter()
+
+            @torch._dynamo.optimize(counter)
+            def f():
+                x = torch.randn(3)
+                return x * 2
+
+            self.assertEqual(f().device.type + ":0", device)
+            self.assertEqual(counter.frame_count, 1)
+
+            torch.set_default_device("cpu")
+
+            self.assertEqual(f().device.type, "cpu")
+            self.assertEqual(counter.frame_count, 2)
+
+        finally:
+            torch.set_default_device(None)
+
+    def test_torch_cuda_is_initialized(self):
+        @torch.compile(fullgraph=True, backend="eager")
+        def f(x):
+            if torch.cuda.is_initialized():
+                return x + 1
+            return x + 2
+
+        inp = torch.randn(3)
+        self.assertEqual(f(inp), inp + 1)
+
+        with mock.patch("torch.cuda.is_initialized", lambda: False):
+            self.assertEqual(f(inp), inp + 2)
+
+    @requires_cuda
+    def test_zero_dim_param_mixed_device_grad(self):
+        class RegressionModel(torch.nn.Module):
+            def __init__(self, a=0, b=0):
+                super().__init__()
+                self.a = torch.nn.Parameter(torch.tensor(a).float())
+                self.b = torch.nn.Parameter(torch.tensor(b).float())
+
+            def forward(self, x):
+                return x * self.a + self.b
+
+        model = RegressionModel()
+        model.forward = torch.compile(
+            model.forward, backend="aot_eager", fullgraph=True
+        )
+        inputs = torch.randn(4, 10).to("cuda")
+        out = model(inputs)
+        out.sum().backward()
+        self.assertIsNotNone(model.a.grad)
+        self.assertIsNotNone(model.b.grad)
+        self.assertEqual(model.a.grad.device, torch.device("cpu"))
+        self.assertEqual(model.b.grad.device, torch.device("cpu"))
+
+    @unittest.skipIf(not TEST_CUDA, "test requires CUDA")
+    def test_cuda_sync(self):
+        def fn(x):
+            y = x + 1
+            torch.cuda.synchronize()
+            return y * 2
+
+        x = torch.ones(2, device="cuda")
+        cnt = torch._dynamo.testing.CompileCounter()
+        opt_fn = torch.compile(fn, backend=cnt)
+        self.assertEqual(fn(x), opt_fn(x))
+        self.assertEqual(cnt.frame_count, 1)
+
+    def test_current_accelerator(self):
+        @torch.compile(backend="eager", fullgraph=True)
+        def fn(x):
+            torch.accelerator.current_accelerator()
+            return x + 1
+
+        self.assertEqual(fn(torch.ones(3)), torch.ones(3) + 1)
 
 
 devices = ("cuda", "hpu", "xpu")
 instantiate_device_type_tests(
     MiscTestsDevice, globals(), only_for=devices, allow_xpu=True
+)
+instantiate_device_type_tests(
+    ReproDeviceRuntimeTests, globals(), only_for=("cuda", "hpu")
 )
 if __name__ == "__main__":
     from torch._dynamo.test_case import run_tests
